bot/pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "bot"
version = "0.1.1"
dependencies = [
  "discord.py",
  "tzdata",
  "poster @ {root:parent:uri}/poster",
]

[tool.hatch.metadata]
allow-direct-references = true
